refactor(#65): item 4 — lib/tinynn{,_cuda,_metal}.rb → lib/toy/ffi/ (path-only move)
Module names UNCHANGED (TinyNN/TinyNNCuda/TinyNNMetal). Mechanical
require_relative sweep across lib/, examples/, prep/, tep_demo/legacy;
Makefile (147 path refs), prep/gen_cuda_mirror.rb rewrite rules
(lib-root 'toy/ffi/tinynn' + engine-depth '../../ffi/tinynn'),
prep/gen_coverage.rb, prep/build_tep_app.sh, living docs.

Note: examples/legacy/* requires were ALREADY unresolvable (Spinel
warns + ignores, pre-existing); kept textually consistent, not fixed.

Gates (a699cf9): verify-mirrors, gate-compute-surface{,-cuda},
train_gate x4 byte-exact, train_cuda_gate x3 byte-identical,
gate-run-log. Serve (f6d5eef): serve_gate + serve_events_gate PASS,
binary 2.09MB.
