+++ examples/gpt2_train.rb
@@ -23,7 +23,7 @@
 # emit-0 → CE=0. (Bit us once; see docs/notes/gpt2-engine-spinel-blocker.md.)
 
 require_relative "../lib/toy"
-require_relative "../lib/tinynn"
+require_relative "../lib/toy/ffi/tinynn"
 require_relative "../lib/toy/llm/engine/gpt2_seq_engine"
 require_relative "../lib/toy/llm/adamw"
