+++ examples/smoke_corpus_loader.rb
@@ -10,7 +10,7 @@
 # schedule shape (warmup → cosine → tail).
 
 require_relative "../lib/toy/models/transformer"
-require_relative "../lib/tinynn"
+require_relative "../lib/toy/ffi/tinynn"
 require_relative "../lib/toy/io/toy_corpus_loader"
 require_relative "../lib/toy/train/toy_lr_schedule"
