+++ examples/smoke_image_loader.rb
@@ -7,7 +7,7 @@
 #   ./examples/smoke_image_loader
 
 require_relative "../lib/toy/models/transformer"
-require_relative "../lib/tinynn"
+require_relative "../lib/toy/ffi/tinynn"
 require_relative "../lib/toy/io/toy_image_loader"
 
 DIR        = ENV["IMG_DIR"]      || "data/vit_smoke"
