+++ lib/toy/run/train_gpt2_metal.rb
@@ -13,7 +13,7 @@
 # ignored → TinyNNMetal/Mat unloaded → emit-0 → CE=0.
 
 require_relative "../../toy"
-require_relative "../../tinynn_metal"
+require_relative "../ffi/tinynn_metal"
 require_relative "../llm/engine/gpt2_seq_engine_metal"
 
 STEPS    = (ENV["STEPS"]    || "5").to_i
