+++ lib/toy_smollm2_ffi_kv.rb
@@ -16,7 +16,7 @@
 require_relative "toy/models/transformer"
 require_relative "toy"
 require_relative "toy/models/toy_smollm2"
-require_relative "tinynn"
+require_relative "toy/ffi/tinynn"
 # NOTE: not requiring "toy_smollm2_loader" here. Requiring it from
 # inside this file triggers a Spinel GC mark crash in decode_step
 # (sp_gc_mark / sp_PtrArray_new_scan) for reasons we haven't fully
